+++ tests/PhpDATest/Command/Strategy/UsageTest.php
@@ -67,12 +67,19 @@ class UsageTest extends \PHPUnit_Framework_TestCase
         $this->output = \Mockery::mock('Symfony\Component\Console\Output\OutputInterface')->shouldIgnoreMissing();
         $this->config = \Mockery::mock('PhpDA\Command\Config');
 
+        $logger = \Mockery::mock('PhpDA\Parser\Logger');
+        $logger->shouldReceive('getEntries')->andReturn(array());
+        $logger->shouldReceive('getEntries')->andReturn(array());
+        $logger->shouldReceive('isEmpty')->andReturn(true);
+
         $formatter = \Mockery::mock('Symfony\Component\Console\Formatter\OutputFormatterInterface');
         $formatter->shouldIgnoreMissing();
 
         $this->output->shouldReceive('getFormatter')->andReturn($formatter);
         $this->output->shouldReceive('writeln');
         $this->analyzer->shouldReceive('getAnalysisCollection')->andReturn($this->collection);
+        $this->analyzer->shouldReceive('getLogger')->andReturn($logger);
+        $this->builder->shouldReceive('setLogEntries')->with(array());
 
         $filePattern = '*.php';
         $source = './src';
@@ -158,10 +165,6 @@ class UsageTest extends \PHPUnit_Framework_TestCase
         );
         $traverser->shouldReceive('bindVisitors')->once()->with($visitor, $visitorOptions);
 
-        $logger = \Mockery::mock('PhpDA\Parser\Logger');
-        $logger->shouldReceive('isEmpty')->once()->andReturn(true);
-
         $this->analyzer->shouldReceive('getNodeTraverser')->once()->andReturn($traverser);
-        $this->analyzer->shouldReceive('getLogger')->once()->andReturn($logger);
     }
 }