+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - BC: Move GraphViz interface from Builder to Writer
 - BC: Refactor and changing data model in Json Writer
 - BC: Rename fqcn accessor and mutator from fqn to fqcn in Location entity
+- BC: Add logEntries mutator to BuilderInterface
 
 ### Added
 - Add .gitattributes